+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v1.1.0
 
 ### Added
+- Added `guidanceText` and `sampleText` fields to `addQuestionCustomization` and added `json`, `questionText`, `requirementText`, `guidanceText`, `sampleText`, `useSampleTextAsDefault` and `required` to `addCustomQuestionInput` [#130]
+- Added `questionCustomizationByVersionedQuestion` resolver [#130]
+- Added `findByCustomizationAndVersionedQuestion` method to `QuestionCustomization` model [#130]
 - Added some opensearch variables for running it locally in docker-compose.yaml [#118]
 - Added opensearch to the docker compose file
 - Added OpenSearch integration for full-text search of re3data repositories
@@ -96,6 +99,7 @@
 - Removed `ioredis` package
 
 ### Fixed
+- Fixed `fetchTemplateData` query in `TemplateCustomization` model because `questionCustomizationHasSampleText` was incorrectly returning true [#130] 
 - In `preparePaginationOptions` function, wrapped each cursorField with `COALESCE` to handle `NULL` values in SQL `CONCAT`, otherwise if any cursorField is NULL, it would just return a null value due to the way `CONCAT` works [#107]
 - Fixed breaking cloning of template. The `addTemplate` was updated to accept a `copyFromVersionedTemplateId` so that we copy from versioned template, section and questions, when it's not a template from the user's org. Otherwise we check for `copyFromTemplateId` to copy/clone from templates, sections and questions, and if neither are present, we continue to create a new record for `templates` table [#1006]
 - Fixed issue with templates not cloning with sections and questions by updating the `addTemplate` mutation to clone from non-versioned template, section and question [#1006]